refactor: consolidate loading dots component
