+++ apps/kyberswap-interface/src/components/SwapForm/SwapActionButton/SwapButtonWithPriceImpact.tsx
@@ -4,8 +4,8 @@ import { Info } from 'react-feather'
 import { useSearchParams } from 'react-router-dom'
 
 import { ButtonPrimary } from 'components/Button'
+import Dots from 'components/Dots'
 import { MouseoverTooltip } from 'components/Tooltip'
-import { Dots } from 'components/swapv2/styleds'
 import { useDegenModeManager } from 'state/user/hooks'
 import { cn } from 'utils/cn'
 import { checkShouldDisableByPriceImpact } from 'utils/priceImpact'
