+++ apps/kyberswap-interface/src/components/SwapForm/SwapModal/ConfirmSwapModalContent.tsx
@@ -8,6 +8,7 @@ import { calculatePriceImpact } from 'services/route/utils'
 
 import { ButtonOutlined, ButtonPrimary } from 'components/Button'
 import { AutoColumn } from 'components/Column'
+import Dots from 'components/Dots'
 import InfoHelper from 'components/InfoHelper'
 import Loader from 'components/Loader'
 import { RowBetween } from 'components/Row'
@@ -25,7 +26,7 @@ import { TransactionErrorContent } from 'components/TransactionConfirmationModal
 import WarningNote from 'components/WarningNote'
 import { calcPercentFilledOrder } from 'components/swapv2/LimitOrder/helpers'
 import { LimitOrderStatus, LimitOrderTab } from 'components/swapv2/LimitOrder/type'
-import { Dots, StyledBalanceMaxMini } from 'components/swapv2/styleds'
+import { StyledBalanceMaxMini } from 'components/swapv2/styleds'
 import { TOKEN_API_URL } from 'constants/env'
 import { APP_PATHS, PAIR_CATEGORY } from 'constants/index'
 import { useActiveWeb3React } from 'hooks'
