+++ apps/kyberswap-interface/src/components/swapv2/TradePrice.tsx
@@ -3,7 +3,8 @@ import { Trans } from '@lingui/macro'
 import { CSSProperties, ReactNode, useState } from 'react'
 import { Repeat } from 'react-feather'
 
-import { Dots, StyledBalanceMaxMini } from 'components/swapv2/styleds'
+import Dots from 'components/Dots'
+import { StyledBalanceMaxMini } from 'components/swapv2/styleds'
 import { cn } from 'utils/cn'
 import { useCurrencyConvertedToNative } from 'utils/dmm'
 import { formatDisplayNumber } from 'utils/numbers'
