+++ apps/kyberswap-interface/src/pages/MyPool/styleds.tsx
@@ -5,9 +5,3 @@ export const Wrapper = ({ children, className, ...rest }: React.HTMLAttributes<H
     {children}
   </div>
 )
-
-export const Dots = ({ children = '', className, ...rest }: React.HTMLAttributes<HTMLSpanElement>) => (
-  <span className={cn('animate-ellipsis', className)} {...rest}>
-    {children}
-  </span>
-)
