+++ apps/kyberswap-interface/src/pages/PoolFinder/index.tsx
@@ -13,13 +13,13 @@ import { FindPoolTabs } from 'components/NavigationTabs'
 import { NarrowPositionCard } from 'components/PositionCard'
 import Row from 'components/Row'
 import CurrencySearchModal from 'components/SearchModal/CurrencySearchModal'
+import Dots from 'components/Dots'
 import { APP_PATHS } from 'constants/index'
 import { NativeCurrencies } from 'constants/tokens'
 import { PairState, usePair } from 'data/Reserves'
 import { useActiveWeb3React } from 'hooks'
 import useTracking, { TRACKING_EVENT_TYPE } from 'hooks/useTracking'
 import AppBody from 'pages/AppBody'
-import { Dots } from 'pages/MyPool/styleds'
 import { usePairAdderByTokens } from 'state/user/hooks'
 import { useTokenBalances } from 'state/wallet/hooks'
 import { StyledInternalLink } from 'theme'
